Move test to a better location
